fix: fix #104 error in stream-range
Condition 4 method did not check the last set of consecutive Q_PTS.

End of range was changed with + 1 to account for last set of consecutive Q_PTS.

To get way initial code was wrong consider the case when we only have Q_PTS points. range will be range(0,0), an empty range.